🎨 Seperate WarpScreen components
- WarpScreen을 구성하는 요소들을 각각의 컴포넌트로 분리
