+++ packages/client/src/widgets/warpScreen/lib/constants.ts
@@ -11,3 +11,10 @@ type Vector3 = [number, number, number];
 export const SPACE_WARP_CAMERA_POSITION: Vector3 = [0, 100000, 0];
 export const SPACE_WARP_CAMERA_UP: Vector3 = [0, 0, 1];
 export const SPACE_WARP_CAMERA_FAR = 90000;
+
+export const BLOOM_INTENSITY = 2;
+export const BLOOM_MIMPAP_BLUR = true;
+export const BLOOM_LUMINANCE_THRESHOLD = 0.55;
+export const BLOOM_LUMINANCE_SMOOTHING = 0;
+
+export const AMBIENT_LIGHT_INTENSITY = 15;
