+++ packages/client/src/widgets/warpScreen/ui/BrightSphere.tsx
@@ -1,0 +1,8 @@
+export default function BrightSphere() {
+	return (
+		<mesh position={[0, 0, 0]}>
+			<sphereGeometry attach="geometry" args={[600, 32, 32]} />
+			<meshStandardMaterial attach="material" color="white" />
+		</mesh>
+	);
+}
